updated exercise popover
