--- apps/gym/src/components/isolated/Record/ExerciseModal/index.ts
@@ -1,1 +1,0 @@
-export * from './ExerciseModal'
