+++ apps/gym/src/components/isolated/Record/ExercisePopover/index.ts
@@ -1,0 +1,1 @@
+export * from './ExercisePopover'
