+++ packages/gymui/package.json
@@ -5,6 +5,7 @@
   "exports": {
     "./Button": "./src/components/Button/Button.tsx",
     "./Modal": "./src/components/Modal/Modal.tsx",
+    "./Popover": "./src/components/Popover/Popover.tsx",
     "./CloseIcon": "./src/components/CloseIcon/CloseIcon.tsx",
     "./Form": "./src/components/Form/Form.tsx",
     "./Skeleton": "./src/components/Skeleton/Skeleton.tsx",
