+++ packages/gymui/src/components/Popover/Popover.module.css
@@ -14,7 +14,9 @@
   flex-direction: column;
   overflow: hidden;
   position: absolute;
+  top: 80%;
   right: 0;
+  z-index: 1000;
 }
 
 .item_container {
